Remove o11yphant and replace with opentelemetry
Make the opentelemetry usages resilient to otel being disabled, too, by injecting @ConfigProperty referencing the enabled flag. This modification introduces OtelAdapter, which is a convenient way to hide any code that is contingent on the state of the opentelemetry extension, and collects related components for use in the rest of the codebase. I've wired in basic trace fields for the gateway ProxyService, and also use of a child client span to wrap the actual upstream request / response, so we can see latency within the gateway itself by subtracting the two.

This eliminates the need to rely on o11yphant, which is too complex for microservice use anyway.

I've also added (commented-out) examples for configuring the application.yaml for honeycomb, or for a local opentelemetry-collector instance. I've tested that both work.