Streaming for large files (without tracing changes) (#59)
* Add better support for large files

This uses the old trick of setting the Response entity to a StreamingOutput instance, instead of a byte array.

To make this work with the CacheHandler, I had to add access to set a supplier for a tee stream to the StreamingOutput interface, which is how we allow caching efficiently. This means the CacheHandler will register its secondary stream as a Supplier<OutputStream> on the StreamingOutput.

optimize imports and use sha256 for reconfig sensing

I'm also moving us away from Mutiny/Vertx web client and use okhttp

This provides a much simpler way to bridge streams directly back to the user without giving up the async call semantics that we had before...though those calls are simplified a bit too, with the new WebClientAdapter. The point of that adapter class is to allow us to optimize the way we make those upstream proxy calls without changing all the head/get/post/put/delete/etc methods.

I'm also adding micrometer ext for memory metrics and minor cleanup

I fixed an ordering error, along with a couple minor failure conditions in the web client adapter (surrounding retries and error conditions), and demoted the logging in the streaming output class.

After adding micrometer and checking memory usage after downloading two complete indy monolith distro tarballs, I can see that memory consumption rises less than 100kb on each request. This may be explained by path timeout caching in the ServiceConfig class, which I added in the last commit to avoid having to compile / examine the path using regex every time.

* re-enable large file test and clean up a bit

* ignore the LGTM warning for the TeeOutputStream

We cannot close this, because we don't manage the OutputStream it wraps, which is the method parameter.

* Fixing potential resource leak identified by LGTM

* reinstate forbidden content-length header