tighten up span usage
refactor(otel): Shorten field names and remove gateway-specific naming, to make this applicable to sidecar and other services that might wish to copy it.

refactor(otel): Add more fields indicating status on things that could fail, or branches in the logic. This should help us filter and find correlations in outlier measurements.